Move create model table to metadata
